+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This file documents all notable changes to this project. The format of this file
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Traits to handle events, common to all the input types.
+
 ## [0.2.0] - 2019-10-30
 
 ### Added
@@ -21,5 +27,6 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Kafka input/output and an example of their usage.
 
+[Unreleased]: https://github.com/petabi/eventio/compare/0.2.0...HEAD
 [0.2.0]: https://github.com/petabi/eventio/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/petabi/eventio/tree/0.1.0